# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beep communities openapi",
    "description": "API documentation for the Communities service",
    "contact": {
      "name": "communities-core@beep.ovh"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/roles/{role_id}": {
      "get": {
        "tags": [
          "roles"
        ],
        "operationId": "get_role",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "description": "Role ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Server is private"
          },
          "404": {
            "description": "Server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "roles"
        ],
        "operationId": "update_role",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "description": "Role ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid permissions"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the server owner"
          },
          "404": {
            "description": "Server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "role"
        ],
        "operationId": "delete_role",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "description": "Role ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the server owner"
          },
          "404": {
            "description": "Server or role not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/roles/{role_id}/members/{member_id}": {
      "post": {
        "tags": [
          "role"
        ],
        "operationId": "assign_role",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "description": "Role ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "description": "Member ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Role assigned successfully to member"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the server owner"
          },
          "404": {
            "description": "Server or role not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "role"
        ],
        "operationId": "unassign_role",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "description": "Role ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "description": "Member ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Role unassigned successfully from member"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the server owner"
          },
          "404": {
            "description": "Server or role not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/servers/{server_id}/roles": {
      "get": {
        "tags": [
          "roles"
        ],
        "operationId": "list_roles_by_server",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_Vec_Role"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Server is private"
          },
          "404": {
            "description": "Server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "roles"
        ],
        "operationId": "create_role",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid role"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateRoleRequest": {
        "type": "object",
        "required": [
          "name",
          "permissions"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PaginatedResponse_Vec_Role": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "object",
                "required": [
                  "id",
                  "server_id",
                  "name",
                  "permissions",
                  "created_at"
                ],
                "properties": {
                  "created_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "id": {
                    "$ref": "#/components/schemas/RoleId"
                  },
                  "name": {
                    "type": "string"
                  },
                  "permissions": {
                    "$ref": "#/components/schemas/Permissions"
                  },
                  "server_id": {
                    "$ref": "#/components/schemas/ServerId"
                  },
                  "updated_at": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time"
                  }
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "$ref": "#/components/schemas/u64"
          }
        }
      },
      "Permissions": {
        "type": "integer",
        "format": "int32"
      },
      "Role": {
        "type": "object",
        "required": [
          "id",
          "server_id",
          "name",
          "permissions",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/RoleId"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "$ref": "#/components/schemas/Permissions"
          },
          "server_id": {
            "$ref": "#/components/schemas/ServerId"
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "RoleId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ServerId": {
        "type": "string",
        "format": "uuid"
      },
...
      "UpdateRoleInput": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RoleId"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "permissions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "u64": {
        "type": "integer",
        "format": "int64",
        "minimum": 0
      }
    }
  }
}